fix: add tests to only use local storage in brower

+++ src/app/shared/services/localStorage.service.ts
@@ -1,21 +1,31 @@
 /* eslint-disable @typescript-eslint/no-unused-expressions */
 /* eslint-disable @typescript-eslint/member-ordering */
 
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { Character } from '@shared/interfaces/data.interface';
 import { ToastrService } from 'ngx-toastr';
 
 const MY_FAVORITES = 'myFavorites';
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class LocalStorageService {
   private charactersFavSubject = new BehaviorSubject<Character[]>([]);
   charactersFav$ = this.charactersFavSubject.asObservable();
+  charactersFav = [];
+  constructor(
+    private toastrSvc: ToastrService,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {
+    this.initStorage();
+  }
 
-  constructor(private toastrSvc: ToastrService) {
-    this.initialStorage();
+  initStorage() {
+    if (this.platformId === 'browser') {
+      localStorage.setItem('mykey', 'localStorage working');
+      this.initialStorage();
+    }
   }
 
   addOrRemoveFavorite(character: Character): void {
@@ -28,34 +38,52 @@ export class LocalStorageService {
   private addToFavorite(character: Character): void {
     try {
       const currentsFav = this.getFavoritesCharacters();
-      localStorage.setItem(MY_FAVORITES, JSON.stringify([...currentsFav, character]));
+      if (this.platformId === 'browser') {
+        localStorage.setItem(
+          MY_FAVORITES,
+          JSON.stringify([...currentsFav, character])
+        );
+      }
+
       this.charactersFavSubject.next([...currentsFav, character]);
-      this.toastrSvc.success(`${character.name} added to favorite`, 'RickAndMortyAPP');
+      this.toastrSvc.success(
+        `${character.name} added to favorite`,
+        'RickAndMortyAPP'
+      );
     } catch (error) {
       console.log('Error saving localStorage', error);
-      this.toastrSvc.error(`Error saving localStorage ${error} `, 'RickAndMortyAPP');
+      this.toastrSvc.error(
+        `Error saving localStorage ${error} `,
+        'RickAndMortyAPP'
+      );
     }
   }
 
   private removeFromFavorite(id: number): void {
     try {
       const currentsFav = this.getFavoritesCharacters();
-      const characters = currentsFav.filter(item => item.id !== id);
-      localStorage.setItem(MY_FAVORITES, JSON.stringify([...characters]));
+      const characters = currentsFav.filter((item) => item.id !== id);
+      if (this.platformId === 'browser') {
+        localStorage.setItem(MY_FAVORITES, JSON.stringify([...characters]));
+      }
       this.charactersFavSubject.next([...characters]);
       this.toastrSvc.warning(`Removed from favorite`, 'RickAndMortyAPP');
     } catch (error) {
       console.log('Error removing localStorage', error);
-      this.toastrSvc.error(`Error removing localStorage ${error} `, 'RickAndMortyAPP');
+      this.toastrSvc.error(
+        `Error removing localStorage ${error} `,
+        'RickAndMortyAPP'
+      );
     }
-
   }
 
   getFavoritesCharacters(): any {
     try {
-      const charactersFav = JSON.parse(localStorage.getItem(MY_FAVORITES));
-      this.charactersFavSubject.next(charactersFav);
-      return charactersFav;
+      if (this.platformId === 'browser') {
+        this.charactersFav = JSON.parse(localStorage.getItem(MY_FAVORITES));
+      }
+      this.charactersFavSubject.next(this.charactersFav);
+      return this.charactersFav;
     } catch (error) {
       console.log('Error getting favorites from localStorage', error);
     }
@@ -63,7 +91,9 @@ export class LocalStorageService {
 
   clearStorage(): void {
     try {
-      localStorage.clear();
+      if (this.platformId === 'browser') {
+        localStorage.clear();
+      }
     } catch (error) {
       console.log('Error cleaning localStorage', error);
     }
@@ -71,7 +101,7 @@ export class LocalStorageService {
 
   private initialStorage(): void {
     const currents = JSON.parse(localStorage.getItem(MY_FAVORITES));
-    if (!currents) {
+    if (this.platformId === 'browser' && !currents) {
       localStorage.setItem(MY_FAVORITES, JSON.stringify([]));
     }
     this.getFavoritesCharacters();
